Replace references of getProductName with getDisplayName
 The
productName is used for the name of the built 'thing' out of
electron-builder, and we want that to reflect the component name of
"Chef Workstation App"; but the displayed screens/menus
) are all providing information about the Chef Workstation
installed product, and not the app - they should read as "Chef
Workstation" instead of the app name.

This update adds a displayName key with a value of "Chef WOrkstation" to
package.json, and  and uses it in place of productName throughout the user-facing portion
of the application.
Signed-off-by: Marc A. Paradise <marc.paradise@gmail.com>