Omnibus inclusion support: split build targets
This splits out the mac/linux/win build targets, and removes
'publish' because we will not be publishing this outside of the
chef workstation package for the time being.

It also changes the output format to zip for Mac and Windows, and
'directory' for linux.  Linux differs to prevent build-time issues
with incorrect GLIBC version when building the electron app on RHEL6.

The chef-workstation omnibus build will run the electron application
build and include the artifacts in the chef-workstation package.

Signed-off-by: Marc A. Paradise <marc.paradise@gmail.com>