Disable automatic update checks.
We will re-enable this once we have converted update checking
to talk to Omnitruck - we're no longer publishing releases to github/
using the built-in autoupdate mechanism.

In the mean time, leaving it enabled will cause an error on starup/update
check because app-update.yml is not included in the distribution. We're disabling update
checks for now to prevent errors.

Signed-off-by: Marc A. Paradise <marc.paradise@gmail.com>