Hardcoding about page links for now
The real change is to add a `RELEASE_NOTES.md` document to Chef
Workstation that we manually update (like we do for ChefDK). Then
we include that into the Chef Workstation package, read it at runtime
and render it from markdown to HTML.

The package details can be parsed from the output of `chef -v` and
rendered to markdown as well.

Signed-off-by: tyler-ball <tball@chef.io>