Stop using isDev when it prevents local testing
We were using isDev too agressively. It meant when trying to test
locally that we could not actually test auto update functionality. Now
it will attempt to look for the installed version of Chef Workstation to
update. If no version is installed we use a dummy version that does not
test the auto update functionality.

Signed-off-by: tyler-ball <tball@chef.io>