[Link Package Details] Opens Electron Window
This is a proposal solution to display the Package Details in an
Electron Window, why? Mainly because our Electro code needs to be
rendered in the context of Electro, if we want to open a plain HTML on a
web browser, we need to code the javascript isolatet from the electron
module.

This approach reads the gem manifest file on disk, parse it and returns
the specific list of gems that are installed in the users workstation.
It is exactly the same output as if you run `chef -v` on the command
line.
```
$ chef -v
Chef Workstation version: 1.0.11
Chef Infra Client version: 15.2.20
Chef InSpec version: 4.10.4
Test Kitchen version: 2.2.5
Foodcritic version: 16.1.1
Cookstyle version: 0.72.0
```

Signed-off-by: Salim Afiune <afiune@chef.io>