Create functions to access external resources
This commit organizes some helper functions to access different paths
during the lifecycle of this application. Some paths needs to reference
external directories like `Resources/assets` outside the `app.asar`
archive. When you see the work `ExternalXYZDir` it is a path that is
outside the archive, if you find `XYZDir` it means that you are
accessing a path inside the archive.

Read more about `asar`, the packaging method we use for this applications,
in the following link:
https://electronjs.org/docs/tutorial/application-packaging

Signed-off-by: Salim Afiune <afiune@chef.io>