Avoid global variables for AppConfig
In the future we need to create a true Store to cache some of this
data but using global variables is not the right way here, why? because
the state is not shared between Dialog windows and processes.

This commit is actually fixing the case when the menu and other parts of
the app doesn't update when you switch the channel.

Signed-off-by: Salim Afiune <afiune@chef.io>