Fix the scrollbar showing up in the about window
When we moved the contents of this window around we made it slightly too large for the window which results in the scroll bar showing. section has a 20px botton margin which was also being applied below the copyright information. We don't need a margin below this content and that margin was resulting in the scroll bar showing. I added a footer class with a 0px margin since it's the last thing on the window and the margin doesn't matter there. I also added a 10px bottom margin (instead of 20px) to main.

Signed-off-by: Tim Smith <tsmith@chef.io>