Do not display errors on bg update checks
This changes the default behavior of `triggerUpdateCheck` to
assume that the update check was not made by the user. This
takes care of various cases in which an automatic update
check is made in the background (or on startup), and failures
are shown as interrupting dialog boxes/popups.

Also Added logging when an error event is received in update check,
made the error title more meaningful, and provided better text
for the "I do not know what the problem is" scenario.

This also updates the behavior of `preferences` to only
trigger a new update check if updates are enabled,
and to flag them as not user requested so that we don't
show a notification popup or an error when someone checks a
preferences checkbox.

Changed the message
Signed-off-by: Marc A. Paradise <marc.paradise@gmail.com>